refactor: Section 클래스의 wrapper class 를 생성하여 구간에 대한 책임을 분리합니다
